Fixes #17811 - Improve error message LDAP fail
