+++ app/models/external_usergroup.rb
@@ -38,7 +38,7 @@ class ExternalUsergroup < ApplicationRecord
 
   def in_auth_source?(source = auth_source)
     errors.add :name, _("is not found in the authentication source") unless source.valid_group?(name)
-  rescue Net::LDAP::LdapError => e
+  rescue Net::LDAP::Error, Net::LDAP::LdapError => e
     errors.add :auth_source_id, _("LDAP error - %{message}") % { :message => e.message }
   end
